修改了发送消息后接受的结果字段类型，修改了分配客服后发送自定义消息

+++ src/main/java/org/cybercaelum/household_management/pojo/dto/MsgSendResultDTO.java
@@ -18,5 +18,5 @@ import lombok.NoArgsConstructor;
 public class MsgSendResultDTO {
     private String serverMsgID;
     private String clientMsgID;
-    private Integer sendTime;
+    private Long sendTime;
 }
